Update on "[sparse] Add i8i8->i32 support for cuSPARSELt"
Summary:

With the release of cuSPARSELt v0.5.0, we now have support for
int8 int8 -> int32 matmul.

This PR adds support for this via out_dtype.

Test Plan:
```
python test/test_sparse_semi_structured.py -k int32
```

Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]

+++ test/test_sparse_semi_structured.py
@@ -244,7 +244,7 @@ class TestSparseSemiStructured(TestCase):
 
             B = torch.rand((128, 128), device=A_sparse.device).to(torch.int8)
 
-            dense_result = torch.mm(A.cpu(), B.t().cpu()).to(device, dtype=torch.float16)
+            dense_result = torch.mm(A.cpu(), B.t().cpu()).to(device, dtype=torch.int32)
             sparse_result = torch._cslt_sparse_mm(A_sparse.compressed_tensor_cusparselt, B.t(), out_dtype=torch.int32)
             assert torch.allclose(dense_result, sparse_result, rtol=1e-3, atol=1e-3)
